common: Add table selector.
