+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Common.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e400a7e43a3de4549a213ea2c9c2423a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
