+++ VisualPinball.Unity/VisualPinball.Unity.Editor/Common/TableSelectorHook.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6079829b01c9cf84abfc02492d9c5e24
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
