+++ VisualPinball.Unity/VisualPinball.Unity/Common/TableSelector.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f64b1655813b43de927b9bf4c33bc2fe
+guid: de12b67118dd83c4a9030d28b04bbe66
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
